Refactor signature validator into the module
