+++ contracts/mocks/core/CoreMock.sol
@@ -8,11 +8,10 @@ import { ISetToken } from "../../core/interfaces/ISetToken.sol";
 contract CoreMock is Core {
     constructor(
         address _transferProxy,
-        address _vault,
-        address _signatureValidator
+        address _vault
     )
         public
-        Core(_transferProxy, _vault, _signatureValidator)
+        Core(_transferProxy, _vault)
     {}
 
     /*
