+++ test/contracts/scenario/issuance_orders/coreIssuanceOrderScenarios.spec.ts
@@ -81,8 +81,13 @@ contract('CoreIssuanceOrder::Scenarios', accounts => {
     vault = await coreWrapper.deployVaultAsync();
     transferProxy = await coreWrapper.deployTransferProxyAsync();
     signatureValidator = await coreWrapper.deploySignatureValidatorAsync();
-    core = await coreWrapper.deployCoreAsync(transferProxy, vault, signatureValidator);
-    issuanceOrderModule = await coreWrapper.deployIssuanceOrderModuleAsync(core, transferProxy, vault);
+    core = await coreWrapper.deployCoreAsync(transferProxy, vault);
+    issuanceOrderModule = await coreWrapper.deployIssuanceOrderModuleAsync(
+      core,
+      transferProxy,
+      vault,
+      signatureValidator
+    );
     await coreWrapper.addModuleAsync(core, issuanceOrderModule.address);
 
     setTokenFactory = await coreWrapper.deploySetTokenFactoryAsync(core.address);
