+++ utils/contract_logs/core.ts
@@ -228,19 +228,6 @@ export function ProtocolFeeChanged(
   };
 }
 
-export function SignatureValidatorChanged(
-  _coreAddress: Address,
-  _signatureValidator: Address,
-): Log {
-  return {
-    event: 'SignatureValidatorChanged',
-    address: _coreAddress,
-    args: {
-      _signatureValidator,
-    },
-  };
-}
-
 /********** Other Log Utilities **********/
 
 export function extractNewSetTokenAddressFromLogs(
